simple handler test

+++ src/test/java/com/rbkmoney/shumaich/converter/PostingBatchDamselToPostingBatchConverterTest.java
@@ -1,7 +1,7 @@
 package com.rbkmoney.shumaich.converter;
 
 import com.rbkmoney.damsel.shumpune.PostingBatch;
-import com.rbkmoney.shumaich.TestData;
+import com.rbkmoney.shumaich.helpers.TestData;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -16,7 +16,7 @@ public class PostingBatchDamselToPostingBatchConverterTest {
         PostingBatch postingBatchDamsel = TestData.postingBatchDamsel();
         var postingBatch = converter.convert(postingBatchDamsel);
 
-        Assert.assertEquals(postingBatchDamsel.id, postingBatch.getId());
+        Assert.assertEquals(postingBatchDamsel.id, postingBatch.getId().longValue());
         Assert.assertEquals(postingBatchDamsel.getPostings().size(), postingBatch.getPostings().size());
     }
