+++ src/test/java/com/rbkmoney/shumaich/converter/PostingDamselToPostingConverterTest.java
@@ -1,7 +1,7 @@
 package com.rbkmoney.shumaich.converter;
 
 import com.rbkmoney.damsel.shumpune.Posting;
-import com.rbkmoney.shumaich.TestData;
+import com.rbkmoney.shumaich.helpers.TestData;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -14,9 +14,9 @@ public class PostingDamselToPostingConverterTest {
         Posting postingDamsel = TestData.postingDamsel();
         var posting = converter.convert(postingDamsel);
 
-        Assert.assertEquals(postingDamsel.amount, posting.getAmount());
-        Assert.assertEquals(postingDamsel.from_id, posting.getFromId());
-        Assert.assertEquals(postingDamsel.to_id, posting.getToId());
+        Assert.assertEquals(postingDamsel.amount, posting.getAmount().longValue());
+        Assert.assertEquals(postingDamsel.from_id, posting.getFromId().longValue());
+        Assert.assertEquals(postingDamsel.to_id, posting.getToId().longValue());
         Assert.assertEquals(postingDamsel.currency_sym_code, posting.getCurrencySymCode());
         Assert.assertEquals(postingDamsel.description, posting.getDescription());
     }
